pyproject.toml
[project]
name = "timescale-diffusion"
version = "0.1.0"
description = "future frame prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"torch",
	"torchvision",
	"torchmetrics",
	"torchinfo",
	"pytorch-lightning",
	"einops",
	"schedulefree",
	"nvidia-dali-cuda120",
	"nvidia-nvimgcodec-cu12",
	"nvidia-nvjpeg-cu12",
	"nvidia-nvjpeg2k-cu12",
	"polars",
	"torch-fidelity",
	"numpy",
	"scikit-learn",
	"av",
	"pims",
	"wandb",
	"pip-system-certs",
	"typing",
	"argparse",
	"toml",
	"tqdm",
	"matplotlib",
	"pandas",
	"ruff",
]

[tool.uv]
dev-dependencies = []
cache-dir = "/cluster/research-groups/wehrwein/home/finn/.uvcache"

[tool.setuptools]
py-modules = ['tspm', 'data', 'vqvae', 'wandb', 'blocks', 'losses', 'configs', 'datasets', 'artifacts', 'condor_logs', 'experiments']

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 160
indent-width = 4

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

docstring-code-format = false
docstring-code-line-length = "dynamic"

